crawler/driver/qqvideo/qqvideo_test.go
package qqvideo

import (
	"github.com/chamzzzzzz/hot/crawler/driver"
	"testing"
)

func TestCrawl(t *testing.T) {
	c := Crawler{}
	if board, err := c.Crawl(); err != nil {
		t.Error(err)
	} else {
		for _, hot := range board.Hots {
			t.Log(hot)
		}
	}
}

func TestCrawlSearch(t *testing.T) {
	c := Crawler{Option: driver.Option{Catalog: Search}}
	if board, err := c.Crawl(); err != nil {
		t.Error(err)
	} else {
		for _, hot := range board.Hots {
			t.Log(hot)
		}
	}
}

func TestCrawlTV(t *testing.T) {
	c := Crawler{Option: driver.Option{Catalog: TV}}
	if board, err := c.Crawl(); err != nil {
		t.Error(err)
	} else {
		for _, hot := range board.Hots {
			t.Log(hot)
		}
	}
}

func TestCrawlVariety(t *testing.T) {
	c := Crawler{Option: driver.Option{Catalog: Variety}}
	if board, err := c.Crawl(); err != nil {
		t.Error(err)
	} else {
		for _, hot := range board.Hots {
			t.Log(hot)
		}
	}
}

func TestCrawlCartoon(t *testing.T) {
	c := Crawler{Option: driver.Option{Catalog: Cartoon}}
	if board, err := c.Crawl(); err != nil {
		t.Error(err)
	} else {
		for _, hot := range board.Hots {
			t.Log(hot)
		}
	}
}

func TestCrawlChild(t *testing.T) {
	c := Crawler{Option: driver.Option{Catalog: Child}}
	if board, err := c.Crawl(); err != nil {
		t.Error(err)
	} else {
		for _, hot := range board.Hots {
			t.Log(hot)
		}
	}
}

func TestCrawlMovie(t *testing.T) {
	c := Crawler{Option: driver.Option{Catalog: Movie}}
	if board, err := c.Crawl(); err != nil {
		t.Error(err)
	} else {
		for _, hot := range board.Hots {
			t.Log(hot)
		}
	}
}

func TestCrawlDoco(t *testing.T) {
	c := Crawler{Option: driver.Option{Catalog: Doco}}
	if board, err := c.Crawl(); err != nil {
		t.Error(err)
	} else {
		for _, hot := range board.Hots {
			t.Log(hot)
		}
	}
}

func TestCrawlGames(t *testing.T) {
	c := Crawler{Option: driver.Option{Catalog: Games}}
	if board, err := c.Crawl(); err != nil {
		t.Error(err)
	} else {
		for _, hot := range board.Hots {
			t.Log(hot)
		}
	}
}

func TestCrawlMusic(t *testing.T) {
	c := Crawler{Option: driver.Option{Catalog: Music}}
	if board, err := c.Crawl(); err != nil {
		t.Error(err)
	} else {
		for _, hot := range board.Hots {
			t.Log(hot)
		}
	}
}
